[RM] Allows custom config for inline delimiters
These changes allow the consuming application to pass in a custom
configuration for parsing Remark-Math segments inline.

The intended use case is that delimiters can indicate either Inline Mode
or Display (Block) Mode, depending on the delimiter used, and custom
delimiters for either of these.

For the particular use case we are applying, we don't want
single-dollar-sign delimited strings to be interpreted as LaTeX at all,
in case users are intending them to indicate regular text, in possible
ignorance of LaTeX. Instead, we want double-dollar-signs to be used to
indicate Inline Mode, and triple-dollar-signs to be used to indicate
Display (Block) Mode. A test has been added for this configuration.